Refactor scraper service and add new scraper for runemetrics and highscores
